requirements.txt
astroid==2.7.3
black==21.9b0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
paho-mqtt==1.5.0
pathspec==0.9.0
pkg_resources==0.0.0
platformdirs==2.3.0
pylint==2.10.2
regex==2021.8.28
requests==2.26.0
supervisor==4.2.2
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
wrapt==1.12.1
